application/api/requirements.txt
Flask
boto3
pytest
requests
Pillow